accident_detection/requirements.txt
# Django
Django==4.2.8
djangorestframework==3.14.0

# OpenCV
opencv-python==4.8.1.78

# Utilities
Pillow==10.2.0
numpy==1.26.4
requests==2.31.0

# Development
python-dotenv==1.0.0

# Optional: PyTorch and YOLO (comment out if installation issues persist)
# torch==2.2.0
# torchvision==0.17.0
# torchaudio==2.2.0
# ultralytics==8.3.0
